refactor tests structure and naming
