+++ tools/tests/e2e_common_waiters.py
@@ -1,12 +1,13 @@
 import time
 
-from e2e_constants import TIMEOUT
 from selenium.common.exceptions import (
     StaleElementReferenceException,
 )
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
+
+TIMEOUT = 30
 
 
 def wait_for(condition_function, *args):
