--- vizro-ai/tests/tests_utils/e2e_constants.py
@@ -1,11 +1,0 @@
-INVALID_PROP_ERROR = "Invalid prop `persisted_props[0]` of value `on` supplied to `t`"
-REACT_NOT_RECOGNIZE_ERROR = "React does not recognize the `%s` prop on a DOM element"
-SCROLL_ZOOM_ERROR = "_scrollZoom"
-REACT_RENDERING_ERROR = "unstable_flushDiscreteUpdates: Cannot flush updates when React is already rendering"
-UNMOUNT_COMPONENTS_ERROR = "React state update on an unmounted component"
-WILLMOUNT_RENAMED_WARNING = "componentWillMount has been renamed"
-WILLRECEIVEPROPS_RENAMED_WARNING = "componentWillReceiveProps has been renamed"
-READPIXELS_WARNING = "GPU stall due to ReadPixels"
-WEBGL_WARNING = "WebGL"  # https://issues.chromium.org/issues/40277080
-
-TIMEOUT = 30
