+++ vizro-core/hatch.toml
@@ -111,11 +111,6 @@ template = "examples"
 DASH_DEBUG = "true"
 VIZRO_LOG_LEVEL = "DEBUG"
 
-[envs.hatch-uv]
-dependencies = [
-  "uv<0.5.10"  # https://github.com/astral-sh/uv/issues/10039
-]
-
 [envs.lower-bounds]
 extra-dependencies = [
   "pydantic==1.10.16",
